KOGITO-5157: Improving Component architecture.
